# src/api/swagger/swagger.yaml
swagger: "2.0"
info:
  description: "BlaynCafe root API"
  version: "0.1.0"
  title: "BlaynCafe API"
basePath: "/api"
tags:
- name: "Student"
- name: "Event"
- name: "Sponsor"
- name: "Admin"
- name: "Image"
consumes:
- "application/json"
produces:
- "application/json"
schemes:
- "http"
paths:
  /student/register:
    post:
      tags:
      - "Student"
      summary: "Register new student"
      operationId: "newStudent"
      parameters: 
      - in: "body"
        name: "student"
        description: "Student information"
        schema:
          $ref: "#/definitions/SubmitStudent"
      responses:
        405:
          description: "Invalid information"
        201:
          description: "Successful registed as new Student"
          schema:
            type: "string"
            description: "A random barcode for Student"
  /student/login:
    get:
      tags:
      - "Student"
      summary: "Student login"
      operationId: "loginStudent"
      parameters:
      - name: "username"
        in: "query"
        description: "The user name for login"
        required: true
        type: "string"
      - name: "password"
        in: "query"
        description: "The password for login in clear text"
        required: true
        type: "string"
      responses:
        401:
          description: "Invalidate email or password"
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/AccessToken"
  /student/logout:
    get:
      tags:
      - "Student"
      summary: "Student logout"
      operationId: "logoutStudent"
      security:
      - StudentSecurity: []
      responses:
        200:
          description: "Logout successful"
  /student/info:
    get:
      tags:
      - "Student"
      summary: "Get current student information"
      security: 
      - StudentSecurity: []
      operationId: "studentUserInfo"
      responses:
        200:
          description: "Student information"
          schema:
            $ref: '#/definitions/ResponseStudent'
  /student/event:
    get:
      tags:
      - "Event"
      - "Student"
      summary: "View student applied events"
      operationId: "studentListEvent"
      parameters:
      - name: "limit"
        in: "query"
        description: "Limit the number of return results"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "Offset of the return results"
        required: false
        type: "integer"
      - name: "eventStatus"
        in: "query"
        description: "Filter the return events by status"
        required: false
        type: "string"
        enum: ['draft', 'under_review', 'awaiting_hoding', 'held']
      security:
      - StudentSecurity: []
      responses:
        405:
          description: "Invalid input"
        401:
          description: "Invalidate access token"
        200:
          description: "Return all events which student applied"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ResponseEvent"
  /student/{id}/barcode:
    get:
      tags:
      - "Admin"
      summary: "Find student with given barcode and checkout"
      parameters:
      - in: "path"
        name: "id"
        description: "Student ID"
        required: true
        type: "string"
      responses:
        200:
          description: "Return student with given barcode"
          schema:
            $ref: "#/definitions/ResponseStudent"
        404:
          description: "Not found any student with given barcode"
  /student/event/join:
    post:
      tags:
      - "Event"
      - "Student"
      summary: "Enroll student to an event"
      operationId: "enrollStudentEvent"
      parameters:
      - name: "eventId"
        in: "query"
        description: "ID of the event which student wants to join"
        required: true
        type: "string"
      responses:
        405:
          description: "Invalid event ID"
        401:
          description: "Invalidate access token"
        200:
          description: "Successful enrolled event"
      security:
      - StudentSecurity: []
  /sponsor/event/{eventId}/cancel:
    post:
      tags:
      - "Event"
      - "Sponsor"
      summary: "Cancel an event which sponsor created before"
      operationId: "cancelEvent"
      parameters:
      - name: "eventId"
        in: "path"
        description: "ID of the event which sponsor wants to cancel"
        required: true
        type: "string"
      security:
      - SponsorSecurity: []
      responses:
        405:
          description: "Invalid event ID or enrollment cannot be cancel"
        401:
          description: "Invalidate access token"
        200:
          description: "Successful enrolled event"
  /sponsor/event:
    get:
      tags:
      - "Event"
      - "Sponsor"
      summary: "View sponsor created events"
      operationId: "sponsorListEvent"
      parameters:
      - name: "limit"
        in: "query"
        description: "Limit the number of return results"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "Offset of the return results"
        required: false
        type: "integer"
      security:
      - SponsorSecurity: []
      responses:
        405:
          description: "Invalid input"
        401:
          description: "Invalid access token"
        200:
          description: "Return all events which sponsor created"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ResponseEvent"
    post:
      tags:
      - "Event"
      - "Sponsor"
      summary: "Submit an event"
      operationId: "createNewEventBySponsor"
      parameters:
      - in: "body"
        name: "event"
        description: "Event's information which sponsor want to create"
        schema:
          type: "object"
          properties:
            event:
              $ref: "#/definitions/SubmitEvent"
      security:
      - SponsorSecurity: []
      responses:
        405:
          description: "Invalid input"
        401:
          description: "Invalid access token"
        201:
          description: "Successful submiited new event"
  /sponsor/event/cancel:
    post:
      tags:
      - "Event"
      - "Sponsor"
      summary: "Cancel an event which sponsor created before"
      parameters:
      - in: "body"
        name: "eventId"
        description: "ID of the event which sponsor wants to cancel"
        required: true
        schema:
          type: "string"
      security:
      - SponsorSecurity: []
      responses:
        405:
          description: "Invalid event ID or enrollment cannot be cancel"
        401:
          description: "Invalidate access token"
        200:
          description: "Successful enrolled event"
  /sponsor/login:
    get:
      tags:
      - "Sponsor"
      summary: "Sponsor login"
      operationId: "loginSponsor"
      parameters:
      - name: "username"
        in: "query"
        description: "The user name for login"
        required: true
        type: "string"
      - name: "password"
        in: "query"
        description: "The password for login in clear text"
        required: true
        type: "string"
      responses:
        401:
          description: "Invalidate email or password"
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/AccessToken"
  /sponsor/logout:
    get:
      tags:
      - "Sponsor"
      operationId: "logoutSponsor"
      summary: "Sponsor logout"
      security:
      - SponsorSecurity: []
      responses:
        200:
          description: "Logout successful"
  /student:
    get:
      tags:
      - "Admin"
      summary: "View list students"
      description: ""
      operationId: "studentList"
      parameters:
      - name: "limit"
        in: "query"
        description: "Limit the number of return results"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "Offset of the return results"
        required: false
        type: "integer"
      responses:
        405:
          description: "Invalid input"
        401:
          description: "Invalidate access token"
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ResponseStudent"
      security:
      - AdminSecurity: []
  /student/{id}/status:
    post:
      tags:
      - "Admin"
      summary: "Set status of student: member, under review, deactivate"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of student to return"
        required: true
        type: "string"
      - in: "body"
        name: "status"
        description: "Status will be assign to student"
        required: true
        schema:
          type: "string"
          enum:
          - "member"
          - "under_review"
          - "deactivate"
      responses:
        200:
          description: "Successful change student status"
        404:
          description: "Not found student with that ID"
      security:
      - AdminSecurity: []
  /student/{id}:
    get:
      tags:
      - "Admin"
      summary: "Get student information"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of student to return"
        required: true
        type: "string"
      responses:
        404:
          description: "Invalid student ID"
        200:
          description: "Return student information with studentId"
          schema:
            $ref: "#/definitions/ResponseStudent"
      security:
      - AdminSecurity: []
  /sponsor:
    get:
      tags:
      - "Admin"
      summary: "View list sponsor"
      description: ""
      operationId: "sponsorList"
      parameters:
      - name: "limit"
        in: "query"
        description: "Limit the number of return results"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "Offset of the return results"
        required: false
        type: "integer"
      responses:
        405:
          description: "Invalid input"
        401:
          description: "Invalidate access token"
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ResponseSponsor'
      security:
      - AdminSecurity: []
  /sponsor/{id}:
    get:
      tags:
      - "Admin"
      summary: "Get sponsor information"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of sponsor to return"
        required: true
        type: "string"
      responses:
        404:
          description: "Invalid sponsor ID"
        200:
          description: "Return sponsor information with that ID"
          schema:
            $ref: '#/definitions/ResponseSponsor'
      security:
      - AdminSecurity: []
  /event/{id}:
    get:
      tags:
      - "Admin"
      - "Student"
      - "Sponsor"
      - "Event"
      summary: "Get event information"
      operationId: "getEventInfo"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of event to return"
        required: true
        type: "string"
      responses:
        404:
          description: "Invalid event ID"
        200:
          description: "Return event information with that ID"
          schema:
            $ref: "#/definitions/ResponseEvent"
      security:
      - AdminSecurity: []
      - SponsorSecurity: []
      - StudentSecurity: []
    put:
      tags:
      - "Admin"
      - "Sponsor"
      summary: "Edit event information"
      operationId: "editEvent"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of event to edit"
        required: true
        type: "string"
      - in: 'body'
        name: 'event'
        description: 'updated event'
        schema:
          type: "object"
          properties:
            event:
              $ref: "#/definitions/SubmitEvent"
      responses:
        404:
          description: "Event with id not found or not created by this sponsor"
        201:
          description: "Successful edit event"
      security: 
      - AdminSecurity: []
      - SponsorSecurity: []
    delete:
      tags:
      - "Admin"
      - "Sponsor"
      summary: "Delete an event"
      operationId: "deleteEvent"
      parameters: 
      - name: "id"
        in: "path"
        description: "ID of event to delete"
        required: true
        type: "string"
      responses:
        404:
          description: "Event with id not found or not created by this sponsor"
        200:
          description: "Successful delete event"
      security: 
      - AdminSecurity: []
      - SponsorSecurity: []
  /event/{id}/student:
    get:
      tags: 
      - "Admin"
      - "Sponsor"
      - "Event"
      summary: "Get list students who have joined the event"
      operationId: "getListStudentJoinEvent"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of event to return"
        required: true
        type: "string"
      responses:
        404:
          description: "Event with id not found or not created by this sponsor"
        200:
          description: "List students who have joined the event"
          schema:
            type: 'array'
            items:
              $ref: '#/definitions/ResponseStudent'
      security: 
      - AdminSecurity: []
      - SponsorSecurity: []
  /slotremain:
    get:
      tags:
      - "Admin"
      - "Student"
      operationId: "getSlotRemain"
      summary: "Get remaining slot in Blayncafe"
      responses:
        200:
          description: "Number of remaining slot in Blayncafe"
          schema:
            type: 'integer'
      security: 
      - AdminSecurity: []
      - StudentSecurity: []
  /event:
    get:
      tags:
      - "Admin"
      - "Student"
      - "Sponsor"
      summary: "Get all events"
      operationId: "listAllEvents"
      parameters:
      - name: "limit"
        in: "query"
        description: "Limit the number of return results"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "Offset of the return results"
        required: false
        type: "integer"
      security: 
      - AdminSecurity: []
      - StudentSecurity: []
      - SponsorSecurity: []
      responses:
        200:
          description: "Return all event"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ResponseEvent'
        
securityDefinitions:
  AdminSecurity:
    type: "apiKey"
    name: "ADMIN-API-KEY"
    in: "header"
  StudentSecurity:
    type: "apiKey"
    name: "STUDENT-API-KEY"
    in: "header"
  SponsorSecurity:
    type: "apiKey"
    name: "SPONSOR-API-KEY"
    in: "header"
definitions:
  BaseSponsor:
    type: "object"
    properties:
      companyName:
        type: "string"
        maxLength: 255
      staffName:
        type: "string"
        maxLength: 255
      staffRubyName:
        type: "string"
        maxLength: 255
      companyAddress:
        type: "string"
      companyPhoneNumber:
        type: "string"
      email:
        type: "string"
        format: "email"
      password:
        type: "string"
        format: "password"
      departure:
        type: "string"
      position:
        type: "string"
  ResponseSponsor:
    allOf:
    - $ref: '#/definitions/BaseSponsor'
    - type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
  SubmitSponsor:
    allOf:
    - $ref: '#/definitions/BaseSponsor'
  BaseStudent:
    type: "object"
    properties:
      name:
        type: "string"
        maxLength: 255
      email:
        type: "string"
        format: "email"
      admissionYear:
        type: "string"
        minimum: 2010
        maximum: 2050
      department:
        type: "string"
      dateOfBirth:
        type: "string"
        format: "date"
      password:
        type: "string"
        format: "password"
      studentCard:
        type: "string"
        format: "url"
      image:
        type: "string"
        format: "url"
      studentNumber:
        type: "string"
  SubmitStudent:
    allOf:
    - $ref: '#/definitions/BaseStudent'
  ResponseStudent:
    allOf: 
    - $ref: '#/definitions/BaseStudent'
    - type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
  BaseEvent:
    type: "object"
    properties:
      eventTitle:
        type: "string"
        maxLength: 1000
      date:
        type: "string"
        format: "date"
      startingTime:
        type: "string"
        format: "date-time"
      recruitmentNumbers:
        type: "integer"
        maximum: 5
        minimum: 1
      images:
        type: "string"
        format: "url"
      status:
        type: "string"
        enum: ['draft','under_review','approved','held']
      description:
        type: "string"
  SubmitEvent:
    allOf: 
    - $ref: '#/definitions/BaseEvent'
  ResponseEvent:
    allOf: 
    - $ref: '#/definitions/BaseEvent'
    - type: 'object'
      properties:
        id:
          type: "integer"
          format: "int64"
        sponsorId:
          type: "integer"
          format: "int64"
  AccessToken:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      accessToken:
        type: "string"
      expireTime:
        type: "integer"
        format: "int64"